+++ packages/react/src/components/DataTable/stories/DataTable-batch-actions.stories.js
@@ -79,6 +79,7 @@ export const Default = () => (
         <TableContainer
           title="DataTable"
           description="With batch actions. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas accumsan mauris sed congue egestas. Integer varius mauris vel arcu pulvinar bibendum non sit amet ligula. Nullam ut nisi eu tellus aliquet vestibulum vel sit amet odio."
+          data-testid="urmom"
           {...getTableContainerProps()}>
           <TableToolbar {...getToolbarProps()}>
             <TableBatchActions {...batchActionProps}>
